chore: break down components
