--- packages/frontend/src/components/Explorer/ExploreTree/TableTree/TableTree.styles.ts
@@ -1,22 +1,0 @@
-import { Colors, Icon } from '@blueprintjs/core';
-import styled from 'styled-components';
-
-export const TooltipContent = styled.p`
-    margin: 0;
-    max-width: 360px;
-`;
-
-export const ItemOptions = styled.div`
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    height: 30px;
-`;
-
-export const Placeholder = styled.div`
-    width: 30px;
-`;
-
-export const WarningIcon = styled(Icon)`
-    color: ${Colors.ORANGE5} !important;
-`;
