+++ packages/frontend/src/components/Explorer/ExploreTree/TableTree/Tree/TreeSingleNode.tsx
@@ -1,0 +1,104 @@
+import { Classes, Colors, Icon, Text } from '@blueprintjs/core';
+import { Tooltip2 } from '@blueprintjs/popover2';
+import {
+    DimensionType,
+    friendlyName,
+    isAdditionalMetric,
+    isDimension,
+    MetricType,
+} from '@lightdash/common';
+import React, { FC } from 'react';
+import { useToggle } from 'react-use';
+import { getItemBgColor } from '../../../../../hooks/useColumns';
+import HighlightedText from '../../../../common/HighlightedText';
+import { Hightlighed, Row } from '../TableTree.styles';
+import CustomMetricButtons from './CustomMetricButtons';
+import FieldButtons from './FieldButtons';
+import { Node, useTableTreeContext } from './TreeProvider';
+
+export const getItemIconName = (type: DimensionType | MetricType) => {
+    switch (type) {
+        case DimensionType.STRING || MetricType.STRING:
+            return 'citation';
+        case DimensionType.NUMBER || MetricType.NUMBER:
+            return 'numerical';
+        case DimensionType.DATE || MetricType.DATE:
+            return 'calendar';
+        case DimensionType.BOOLEAN || MetricType.BOOLEAN:
+            return 'segmented-control';
+        case DimensionType.TIMESTAMP:
+            return 'time';
+        default:
+            return 'numerical';
+    }
+};
+
+const TreeSingleNode: FC<{ node: Node; depth: number }> = ({ node, depth }) => {
+    const [isHover, toggle] = useToggle(false);
+    const {
+        itemsMap,
+        selectedItems,
+        isSearching,
+        searchResults,
+        searchQuery,
+        onItemClick,
+    } = useTableTreeContext();
+    const isSelected = selectedItems.has(node.key);
+    const isVisible = !isSearching || searchResults.has(node.key);
+    const item = itemsMap[node.key];
+
+    if (!item || !isVisible) {
+        return null;
+    }
+
+    const label: string =
+        isDimension(item) && item.group
+            ? friendlyName(item.name.replace(item?.group, ''))
+            : item.label || item.name;
+    return (
+        <Row
+            depth={depth}
+            selected={isSelected}
+            bgColor={getItemBgColor(item)}
+            onClick={() => onItemClick(node.key, item)}
+            onMouseEnter={() => toggle(true)}
+            onMouseLeave={() => toggle(false)}
+        >
+            <Icon
+                icon={getItemIconName(item.type)}
+                color={isDimension(item) ? Colors.BLUE1 : Colors.ORANGE1}
+                size={16}
+                style={{ marginRight: 8 }}
+            />
+            <Tooltip2
+                content={item.description}
+                className={Classes.TEXT_OVERFLOW_ELLIPSIS}
+            >
+                <Text ellipsize>
+                    <HighlightedText
+                        text={label}
+                        query={searchQuery || ''}
+                        highlightElement={Hightlighed}
+                    />
+                </Text>
+            </Tooltip2>
+            <span style={{ flex: 1 }} />
+            {isAdditionalMetric(item) ? (
+                <CustomMetricButtons
+                    node={item}
+                    isHovered={isHover}
+                    isSelected={isSelected}
+                />
+            ) : (
+                <FieldButtons
+                    node={item}
+                    onOpenSourceDialog={() => undefined}
+                    isHovered={isHover}
+                    isSelected={isSelected}
+                />
+            )}
+        </Row>
+    );
+};
+
+export default TreeSingleNode;
